trimming and re-organizing to encapsulate borrowed API
