+++ SPISearch/SPISearchAPI.swift
@@ -6,9 +6,6 @@
 //
 
 import Foundation
-//import PackageListTool // for SwiftPackageIndexAPI
-
-import SPIAPI
 
 enum SPIBaseURLs: String {
     case localhost = "http://localhost"
@@ -22,9 +19,6 @@ enum SPIBaseURLs: String {
 // PROD: https://redocly.github.io/redoc/?url=https://swiftpackageindex.com/openapi/openapi.json
 
 enum APISearchParser {
-    static func getSearch(terms: String) async throws {
-        try await SPIAPI.search(terms: terms)
-    }
     
     static func recordSearch(terms: String, from: SPIBaseURLs = .dev) async -> RecordedSearchResult {
         let apiEndpoint = SwiftPackageIndexAPI(baseURL: from.rawValue, apiToken: "")
